commit-24

+++ API/SignalR/MessageHub.cs
@@ -18,16 +18,17 @@ namespace API.SignalR {
 
             var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-            await AddToGroup(groupName);
-
+            var group = await AddToGroup(groupName);
+            await Clients.Group(groupName).SendAsync("UpdateGroup", group);
             var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
 
-            await Clients.Group(groupName).SendAsync("ReceivedMessageThread", messages);
+            await Clients.Caller.SendAsync("ReceivedMessageThread", messages);
         }
 
         public override async Task OnDisconnectedAsync(Exception? exception) {
-
-            await RemoveFromMessageGroup();
+            
+            var group = await RemoveFromMessageGroup();
+            await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
             await base.OnDisconnectedAsync(exception);
         }
 
@@ -71,7 +72,7 @@ namespace API.SignalR {
 
         }
 
-        private async Task<bool> AddToGroup(string groupName) {
+        private async Task<Group> AddToGroup(string groupName) {
 
             var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
 
@@ -84,15 +85,19 @@ namespace API.SignalR {
             } 
 
             group.Connections.Add(connection);
-            return await messageRepository.SaveAllAsync();
+            if(await messageRepository.SaveAllAsync()) return group;
+            throw new HubException("Failed to join group");
         }
 
-        private async Task RemoveFromMessageGroup() {
-            var connection = await messageRepository.GetConnection(Context.ConnectionId);
-            if(connection != null) {
+        private async Task<Group> RemoveFromMessageGroup() {
+            var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+            var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+            if(connection != null && group != null) {
                 messageRepository.RemoveConnection(connection);
-                await messageRepository.SaveAllAsync();
+                if(await messageRepository.SaveAllAsync()) return group;
             }
+
+            throw new Exception("Failed to remove from group");
         }
 
         private string GetGroupName(string caller, string? other) {
